Gather ufunc tests
