+++ py-polars/tests/unit/operations/map/test_map_batches.py
@@ -77,13 +77,3 @@ def test_map_deprecated() -> None:
         pl.col("a").map(lambda x: x)
     with pytest.deprecated_call():
         pl.LazyFrame({"a": [1, 2]}).map(lambda x: x)
-
-
-def test_ufunc_recognition() -> None:
-    df = pl.DataFrame({"a": [1, 1, 2, 2], "b": [1.1, 2.2, 3.3, 4.4]})
-    assert_frame_equal(df.select(np.exp(pl.col("b"))), df.select(pl.col("b").exp()))
-
-
-def test_grouped_ufunc() -> None:
-    df = pl.DataFrame({"id": ["a", "a", "b", "b"], "values": [0.1, 0.1, -0.1, -0.1]})
-    df.group_by("id").agg(pl.col("values").log1p().sum().pipe(np.expm1))
